implement test_scalar_compat
