+++ pandas/tests/indexes/datetimes/test_datetime.py
@@ -2,7 +2,7 @@
 import pytest
 
 import numpy as np
-from datetime import date, timedelta, time, datetime
+from datetime import date, timedelta, time
 
 import dateutil
 import pandas as pd
@@ -14,31 +14,6 @@ from pandas import (DatetimeIndex, Index, date_range, DataFrame,
 from pandas.util.testing import assert_almost_equal
 
 randn = np.random.randn
-
-
-class TestDatetimeIndexLikeTimestamp(object):
-    # Tests for DatetimeIndex behaving like a vectorized Timestamp
-
-    def test_dti_date_out_of_range(self):
-        # see gh-1475
-        pytest.raises(ValueError, DatetimeIndex, ['1400-01-01'])
-        pytest.raises(ValueError, DatetimeIndex, [datetime(1400, 1, 1)])
-
-    def test_timestamp_fields(self):
-        # extra fields from DatetimeIndex like quarter and week
-        idx = tm.makeDateIndex(100)
-
-        fields = ['dayofweek', 'dayofyear', 'week', 'weekofyear', 'quarter',
-                  'days_in_month', 'is_month_start', 'is_month_end',
-                  'is_quarter_start', 'is_quarter_end', 'is_year_start',
-                  'is_year_end', 'weekday_name']
-        for f in fields:
-            expected = getattr(idx, f)[-1]
-            result = getattr(Timestamp(idx[-1]), f)
-            assert result == expected
-
-        assert idx.freq == Timestamp(idx[-1], idx.freq).freq
-        assert idx.freqstr == Timestamp(idx[-1], idx.freq).freqstr
 
 
 class TestDatetimeIndex(object):
@@ -370,18 +345,6 @@ class TestDatetimeIndex(object):
 
         assert_almost_equal(index.isin([index[2], 5]),
                             np.array([False, False, True, False]))
-
-    def test_time(self):
-        rng = pd.date_range('1/1/2000', freq='12min', periods=10)
-        result = pd.Index(rng).time
-        expected = [t.time() for t in rng]
-        assert (result == expected).all()
-
-    def test_date(self):
-        rng = pd.date_range('1/1/2000', freq='12H', periods=10)
-        result = pd.Index(rng).date
-        expected = [t.date() for t in rng]
-        assert (result == expected).all()
 
     def test_does_not_convert_mixed_integer(self):
         df = tm.makeCustomDataframe(10, 10,
