+++ pandas/tests/indexes/datetimes/test_datetimelike.py
@@ -1,9 +1,7 @@
 """ generic tests from the Datetimelike class """
 
-import numpy as np
-import pandas as pd
 from pandas.util import testing as tm
-from pandas import Series, Index, DatetimeIndex, date_range
+from pandas import DatetimeIndex, date_range
 
 from ..datetimelike import DatetimeLike
 
@@ -27,31 +25,7 @@ class TestDatetimeIndex(DatetimeLike):
         pass
 
     def test_intersection(self):
-        first = self.index
-        second = self.index[5:]
-        intersect = first.intersection(second)
-        assert tm.equalContents(intersect, second)
-
-        # GH 10149
-        cases = [klass(second.values) for klass in [np.array, Series, list]]
-        for case in cases:
-            result = first.intersection(case)
-            assert tm.equalContents(result, second)
-
-        third = Index(['a', 'b', 'c'])
-        result = first.intersection(third)
-        expected = pd.Index([], dtype=object)
-        tm.assert_index_equal(result, expected)
+        pass  # handled in test_setops
 
     def test_union(self):
-        first = self.index[:5]
-        second = self.index[5:]
-        everything = self.index
-        union = first.union(second)
-        assert tm.equalContents(union, everything)
-
-        # GH 10149
-        cases = [klass(second.values) for klass in [np.array, Series, list]]
-        for case in cases:
-            result = first.union(case)
-            assert tm.equalContents(result, everything)
+        pass  # handled in test_setops
